refactor: remove individual DI containers for flows
